A lot of stuff. Adding an ID property to Animations, Materials, Scenes, and Meshes to ease retrieval. Library.Meshes, Animations, Materials, and Worlds is now a slice instead of a map. This is to make it consisten to obtain resources from ID or name. Adding Library functions to get elements by ID and name. Optimizing bounds checking for sphere <> triangles and capsule <> triangles by using squared distances when possible. Node.WorldRotation() / Node.WorldScale() are now faster by not decomposing unless it's necessary (which also contributes to bounds checks and raycasts being faster). Broadphase checks no longer use a working set for determining which triangles to check (at possibly the cost of returning slightly too many triangles). Ray casting is also slightly faster on top of the above speed improvements due to getting triangles in a range rather than doing a spherical bounds collision check. All of these changes should mean raycasting is about 50% faster, IIRC. Adding Vector2 type as a game object property type. Adding node count to camera's debug info. Dimensions.Inside(Vector3) > Vector3.IsInsideDimensions(Dimensions). Adding VertexSelection.ScaleUVsVec() and SetUVOffsetVec(). Fixing VertexSelection.Count() to work with fully selected meshes. Renaming Node.InSceneTree() > IsInSceneTree(). Altering Node.HierarchyAsString() to add in the node's iD in its result. Node.SearchTree() now returns a NodeFilter directly rather than a pointer to one. Re-fixing NodeFilter.ByParentProp() to take multiple properties. Adding NodeFilter.ByID() and ByNamePartial() (if you don't need the regex). Adding Property.IsVector2() / AsVector2(). Adding INode.SetWorldScale() (adding it to the interface). Adding Matrix4.SetFloats([]float32), Round(unitSize), ToFloats(), ToFloatSlice(), RowAsVector3(). Exposing Matrix4.setIndex() as Matrix4.SetByIndex(). Renaming Matrix4.ScaleByScalar() > Scale(). Formatting for empty object references in Blender (when drawing game properties to the screen) slightly changed from "< No Object >" to "{ No Object }" Adding Vector2/3.Clone() (which doesn't do anything for Go, but can be useful when working with interpreted languages with Go interop). Adding Vector3.ClampToDimensions() / IsInsideDimensions(), Vector2.Swizzle3 / Swizzle2, Vector2.IsNaN(), IsInf(), Floor(), Ceil().

+++ math32/math32.go
@@ -494,4 +494,4 @@ func Trunc(x float32) float32 {
 	return float32(math.Trunc(float64(x)))
 }
 
-const Pi = math.Pi
+const Pi = float32(math.Pi)
